refactor(package): flatten litestar_mcp internal layout
Ch5 of v0.5.0-consumer-readiness.

Merge:
- filters.py, decorators.py, _descriptions.py, _uri_template.py -> utils.py
- auth/_cache.py, auth/_oidc.py -> auth/oidc.py
- auth/config.py -> auth/backend.py

Public package surface (from litestar_mcp import ... / from
litestar_mcp.auth import ...) is unchanged. Internal callers and the test
suite re-point their imports to the new module paths. No compat shims —
direct imports from the deleted submodules (from litestar_mcp.decorators
import mcp_tool etc.) no longer resolve.

465 tests green; coverage ≥80%.
