+++ docs/examples/snippets/marking_decorator.py
@@ -3,7 +3,7 @@
 from litestar import Litestar, get
 
 from litestar_mcp import LitestarMCP
-from litestar_mcp.decorators import mcp_tool
+from litestar_mcp.utils import mcp_tool
 
 
 def build() -> Litestar:
