+++ litestar_mcp/__init__.py
@@ -21,9 +21,9 @@ from litestar_mcp.auth import (
     create_oidc_validator,
 )
 from litestar_mcp.config import MCPConfig, MCPOptKeys
-from litestar_mcp.decorators import mcp_resource, mcp_tool
 from litestar_mcp.plugin import LitestarMCP
 from litestar_mcp.routes import MCPController
+from litestar_mcp.utils import mcp_resource, mcp_tool
 
 __all__ = (
     "DefaultJWKSCache",
