+++ litestar_mcp/auth/backend.py
@@ -1,9 +1,18 @@
-"""MCPAuthBackend — the built-in bearer/OIDC authentication middleware for MCP."""
+"""MCPAuthBackend + auth configuration dataclasses.
+
+This module consolidates the built-in bearer/OIDC authentication
+middleware (:class:`MCPAuthBackend`) with the configuration dataclasses
+that describe OIDC providers (:class:`OIDCProviderConfig`) and the
+protected-resource discovery manifest (:class:`MCPAuthConfig`). Before
+v0.5.0 these lived in separate modules; Ch5 of the v0.5.0 roadmap
+flattens them here.
+"""
 
 from __future__ import annotations
 
 import inspect
 from collections.abc import Awaitable, Callable, Sequence
+from dataclasses import dataclass, field
 from typing import TYPE_CHECKING, Any
 
 from litestar.exceptions import NotAuthorizedException
@@ -12,15 +21,25 @@ from litestar.middleware.authentication import (
     AuthenticationResult,
 )
 
-from litestar_mcp.auth._oidc import _validate_with_oidc_provider
+from litestar_mcp.auth.oidc import (
+    DEFAULT_CLOCK_SKEW_SECONDS,
+    DEFAULT_JWKS_CACHE_TTL_SECONDS,
+    _validate_with_oidc_provider,
+)
 
 if TYPE_CHECKING:
     from litestar.connection import ASGIConnection
     from litestar.types import ASGIApp, Method, Scopes
 
-    from litestar_mcp.auth.config import OIDCProviderConfig
+    from litestar_mcp.auth.oidc import JWKSCache
 
-__all__ = ("MCPAuthBackend", "TokenValidatorFn", "UserResolver")
+__all__ = (
+    "MCPAuthBackend",
+    "MCPAuthConfig",
+    "OIDCProviderConfig",
+    "TokenValidatorFn",
+    "UserResolver",
+)
 
 UserResolver = Callable[["dict[str, Any]", Any], "Awaitable[Any] | Any"]
 """Callable ``(claims, app) -> user``; sync or async."""
@@ -31,6 +50,63 @@ TokenValidatorFn = Callable[[str], "Awaitable[dict[str, Any] | None]"]
 _BEARER_PREFIX = "Bearer "
 _MISSING_HEADER_MSG = "Missing or invalid Authorization header"
 _INVALID_TOKEN_MSG = "Invalid token"  # noqa: S105 — auth failure message, not a credential
+
+
+# ---------------------------------------------------------------------------
+# Configuration dataclasses
+# ---------------------------------------------------------------------------
+
+
+@dataclass
+class OIDCProviderConfig:
+    """Configuration for validating bearer tokens against an OIDC/JWKS provider.
+
+    Attributes:
+        issuer: Expected ``iss`` claim and discovery base URL.
+        audience: Expected ``aud`` claim (string, list, or ``None`` to skip).
+        jwks_uri: Optional explicit JWKS endpoint (overrides discovery).
+        discovery_url: Optional override for the OpenID discovery document URL.
+        algorithms: Allowed JWS algorithms (default: ``["RS256"]``).
+        cache_ttl: JWKS / discovery document cache TTL in seconds.
+        clock_skew: Tolerance in seconds for ``exp`` / ``iat`` / ``nbf`` checks.
+        jwks_cache: Optional shared :class:`~litestar_mcp.auth.JWKSCache`
+            instance. When ``None`` the process-wide default cache is used.
+    """
+
+    issuer: str
+    audience: str | list[str] | None = None
+    jwks_uri: str | None = None
+    discovery_url: str | None = None
+    algorithms: list[str] = field(default_factory=lambda: ["RS256"])
+    cache_ttl: int = DEFAULT_JWKS_CACHE_TTL_SECONDS
+    clock_skew: int = DEFAULT_CLOCK_SKEW_SECONDS
+    jwks_cache: JWKSCache | None = None
+
+
+@dataclass
+class MCPAuthConfig:
+    """Metadata for the ``/.well-known/oauth-protected-resource`` manifest.
+
+    Authentication *enforcement* is handled by a Litestar authentication
+    middleware (either your own
+    :class:`~litestar.middleware.authentication.AbstractAuthenticationMiddleware`
+    subclass or the built-in :class:`MCPAuthBackend`). This struct only
+    describes the auth surface to discovery clients.
+
+    Attributes:
+        issuer: OAuth 2.1 authorization server issuer URL (advertised to clients).
+        audience: Resource identifier used in the protected-resource manifest.
+        scopes: Mapping of scope name to human-readable description.
+    """
+
+    issuer: str | None = None
+    audience: str | list[str] | None = None
+    scopes: dict[str, str] | None = None
+
+
+# ---------------------------------------------------------------------------
+# Authentication middleware
+# ---------------------------------------------------------------------------
 
 
 class MCPAuthBackend(AbstractAuthenticationMiddleware):
