+++ tests/integration/test_mcp_auth_middleware.py
@@ -19,7 +19,7 @@ from litestar.testing import TestClient
 
 from litestar_mcp import LitestarMCP, MCPConfig
 from litestar_mcp.auth import MCPAuthBackend, MCPAuthConfig
-from litestar_mcp.decorators import mcp_tool
+from litestar_mcp.utils import mcp_tool
 from tests.integration._auth import (
     AUDIENCE,
     FORGED_TOKEN,
