+++ tests/unit/test_discovery.py
@@ -4,7 +4,7 @@ from litestar import Litestar, get
 from litestar.testing import TestClient
 
 from litestar_mcp import LitestarMCP
-from litestar_mcp.decorators import mcp_tool
+from litestar_mcp.utils import mcp_tool
 
 
 def _make_discovery_app() -> Litestar:
