+++ tests/unit/test_tasks.py
@@ -8,7 +8,7 @@ from litestar import Litestar, get
 from litestar.testing import TestClient
 
 from litestar_mcp import LitestarMCP, MCPConfig
-from litestar_mcp.decorators import mcp_tool
+from litestar_mcp.utils import mcp_tool
 
 
 def _ensure_session(client: TestClient[Any]) -> str:
